feat: add sunset/sunrise time, adust backend
This commit accesses the data passed from the backend and adds the
sunset/sunrise times to the hero card. Additionally, to make the data
sent to the frontend more consistent, the data sent back to frontend
uses another get request after storing the data from the api into the
db.