feat: implement seven day forecast, rework hero
This commit implements the weather API into the forecast component,
showing the weather and temperature conditions for the next 7 days. To
make this work, the API requests needed to be changed, forcing a rework
in the hero component. There could be some simplification added to these
components to achieve a more DRY page.