feat: page now renders at start, adjust state
This commit implements the functionality to allow the page to fetchData
at the start, the search box is also switched from using state for
saving the search input to ref, allowing state to not be rerenders on
every keystroke. Framer motion is adjusted to render on each search and
a loading button is rendered while data is being fetched.

frontend